Adds skeleton project with dummy streaming job
This adds a skeleton project which separates the how properties are
supplied from code that uses it. For example, the core library has
no spring dependencies, or other property management utilities.

This borrows the maven configuration from zipkin-dependencies so that
it can be used in spark clusters. Note this isn't fully tested, rather
a scratch pad for others to contribute to.