Adds Adjuster and simplifies other names. Fleshes out README
This adds Adjuster which is the first notable design win in comparison
with the typical Zipkin collector. This accepts an iterable of spans
grouped by trace ID and occurs prior to storage. Multiple adjusters
can be created.

Thanks to @mansu and @naoman for the design idea.

This also simplifies type naming, particularly as it makes the README
easier to write and package and artifact names less long.