Switches serialization approach of storage components
Before, we serialized a pojo to manually create storage components when
in a spark job. This added maintenance as there are configuration items
such an approach will miss, such as okhttp auth interceptors. Now, we
serialize zipkin input properties and use spring to more realistically
reconstruct the storage component.

Fixes #45